pyproject.toml
[tool.poetry]
name = "wise-api"
version = "0.1.0"
description = "Python API client for the Wise API"
authors = ["Jerome Leclanche <jerome@leclan.ch>"]
license = "MIT"
readme = "README.md"
packages = [{include = "wise_api"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pycryptodome = "^3.16.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.991"
isort = "^5.10.1"
types-requests = "^2.28.11.5"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
